pyproject.toml
[tool.poetry]
name = "fastapi-proj"
version = "0.1.0"
description = ""
authors = ["Pavel <pashapasha-2013@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.114.2"
uvicorn = "^0.30.6"
sqlalchemy = "^2.0.34"
aiosqlite = "^0.20.0"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
python-dotenv = "^1.0.1"
fastapi-users = {extras = ["sqlalchemy"], version = "^13.0.0"}
aiogram = "^3.13.0"
requests = "^2.32.3"
redis = "^5.0.8"
apscheduler = "^3.10.4"
aioredis = "^2.0.1"
binary = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
